API: `save` is not called `to_matrix` and a pipeline function; also, it does no longer add arbitrary extensions to the file; closes #61
