+++ SurfaceTopography/IO/common.py
@@ -22,7 +22,86 @@
 # SOFTWARE.
 #
 
-from io import TextIOBase
+import io
+
+###
+
+height_units = {'m': 1.0, 'mm': 1e-3, 'µm': 1e-6, 'nm': 1e-9, 'Å': 1e-10}
+voltage_units = {'kV': 1000.0, 'V': 1.0, 'mV': 1e-3, 'µV': 1e-6, 'nV': 1e-9}
+
+units = dict(height=height_units, voltage=voltage_units)
+
+CHANNEL_NAME_INFO_KEY = 'channel_name'
+
+###
+
+def get_unit_conversion_factor(unit1_str, unit2_str):
+    """
+    Compute factor for conversion from unit1 to unit2. Return None if units are
+    incompatible.
+    """
+    if unit1_str == unit2_str:
+        return 1
+    unit1_kind = None
+    unit2_kind = None
+    unit_scales = None
+    for key, values in units.items():
+        if unit1_str in values:
+            unit1_kind = key
+            unit_scales = values
+        if unit2_str in values:
+            unit2_kind = key
+            unit_scales = values
+    if unit1_kind is None or unit2_kind is None or unit1_kind != unit2_kind:
+        return None
+    return unit_scales[unit1_str] / unit_scales[unit2_str]
+
+
+def mangle_height_unit(unit):
+    unit = unit.strip()
+    if unit == '':
+        return None
+    elif unit == 'A':
+        return 'Å'
+    elif unit == 'μm' or unit == 'um' or unit == '~m':
+        return 'µm'
+    else:
+        return unit
+
+
+###
+
+def is_binary_stream(fobj):
+    """Check whether fobj is a binary stream"""
+    return isinstance(fobj, io.BytesIO) or (
+            hasattr(fobj, 'mode') and 'b' in fobj.mode)
+
+
+def text(func):
+    """Decorator that turns the first argument into a binary stream"""
+
+    def func_wrapper(fobj, *args, **kwargs):
+        close_file = False
+        if not hasattr(fobj, 'read'):
+            fobj = open(fobj, 'r', encoding='utf-8')
+            fobj_text = fobj
+            close_file = True
+        elif is_binary_stream(fobj):
+            fobj_text = io.TextIOWrapper(fobj, encoding='utf-8')
+        else:
+            fobj_text = fobj
+
+        try:
+            retvals = func(fobj_text, *args, **kwargs)
+        finally:
+            if is_binary_stream(fobj):
+                fobj_text.detach()
+                fobj_text = fobj
+            if close_file:
+                fobj_text.close()
+        return retvals
+
+    return func_wrapper
 
 
 class OpenFromAny(object):
@@ -62,13 +141,13 @@ class OpenFromAny(object):
             self._already_open = True
             if self._mode == 'rb':
                 # Turn this into a binary stream, if it is a text stream
-                if isinstance(self._fobj, TextIOBase):
+                if isinstance(self._fobj, io.TextIOBase):
                     # file was opened without the 'b' option, so read its buffer to get the binary data
                     self._fstream = self._fobj.buffer
                 else:
                     self._fstream = self._fobj
             elif self._mode == 'r':
-                if isinstance(self._fobj, TextIOBase):
+                if isinstance(self._fobj, io.TextIOBase):
                     # file was opened without the 'b' option, so read its buffer to get the binary data
                     self._fstream = self._fobj
                 else:
